fix: Ajuste de dockerFile para despliegue en Render
Refactorizacion.

+++ src/AgroVisionPWA/Domain/CoffeePestDetection.Domain/Entities/Farm.cs
@@ -16,7 +16,7 @@ namespace CoffeePestDetection.Domain.Entities
 
         public string? Location { get; set; }
 
-        public double? AreaHectares { get; set; }
+        public decimal? AreaHectares { get; set; }
 
         public bool IsActive { get; set; } = true;
         /** FK */
